Make template insert into highlight
