+++ addon/components/suggested-templates-card.js
@@ -10,19 +10,4 @@ export default class SuggestedTemplatesCardComponent extends Component {
     return this.args.controller;
   }
 
-  get plugin() {
-    return this.args.plugin;
-  }
-
-  @action
-  async insert(template) {
-    await template.reload();
-    const insertedRange = this.controller.executeCommand(
-      'insert-html',
-      instantiateUuids(template.body)
-    );
-    // TODO not the cleanest, we should have more expressive events
-    // so we can simply listen for the contentChange the above insert generates
-    this.plugin.highlightInRange(insertedRange);
-  }
 }
