+++ addon/components/template-provider.hbs
@@ -2,4 +2,5 @@
           applicableTemplates=this.applicableTemplates
           hasApplicableTemplates=this.hasApplicableTemplates
           busy=this.busy
+          insert=this.insert
         )}}
